requirements.txt
moderngl~=5.10.0
numpy~=2.0.1
pygame~=2.6.0
pyglm
easygui~=0.98.3